Background the validation & reading of files during import.
The import process goes through a lengthy process of validation and splitting up files if necessary. This can be done in multiple threads and is implemented as such; however a lot of the code has both UI updates and data extraction making it difficult to do it in a clean way. There is more to be separated from the processing and the UI updates. Saving is not able to be be threaded at this time because of the lexer, which uses globals and this cannot be shared across threads. I have not tested the splitting up of "archive" files which would be something akin to the TrainingCenter history file format; this causes additional files to be generated and added to the end of the queue. I proactively put a mutex to limit inserts to a singular thread during the splitting, something to consider would be a concurrent queue but likely implemented in the same approach.